requirements.txt
gradio==4.44.0
ultralytics==8.3.0
pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
torch==2.4.0
torchvision==0.19.0
